requirements.txt
Django==3.0.7
djangorestframework==3.11.0
pytest
pytest-django
pytest-cov